Rename some things

+++ optimized/reduce.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-uint16_t montgomery_reduce(uint32_t a);
+uint16_t montgomery_reduce(uint64_t a);
 
 #endif
